Create a file for PC design separating from mobile devices
